Moved lens to lenses, service to services
